テストプログラムの整理
