+++ document/error_types.class.pu
@@ -4,37 +4,38 @@ namespace pdfparser::error_types {
 enum error_code <<Enumeration>> {
 	<<Const>>
 	+{static} failed_to_seek_forward_head_of_line             : int = 0
-	+{static} keyword_EOF_not_found                           : int = 1
-	+{static} xref_byte_offset_not_found                      : int = 2
-	+{static} keyword_startxref_not_found                     : int = 3
-	+{static} EOL_not_found                                   : int = 4
-	+{static} signed_integer_not_found                        : int = 5
-	+{static} unsigned_integer_not_found                      : int = 6
-	+{static} keyword_xref_not_found                          : int = 7
-	+{static} space_not_found                                 : int = 8
-	+{static} xref_entry_first_10_digits_invalid              : int = 9
-	+{static} xref_entry_second_5_digits_invalid              : int = 10
-	+{static} xref_entry_keyword_invalid                      : int = 11
-	+{static} xref_entry_last_2_bytes_invalid                 : int = 12
-	+{static} keyword_trailer_not_found                       : int = 13
-	+{static} literal_string_lack_of_right_parenthesis        : int = 14
-	+{static} unknown_character_detected                      : int = 15
-	+{static} object_not_found                                : int = 16
-	+{static} name_invalid_octal_code                         : int = 17
-	+{static} array_lack_of_right_square_bracket              : int = 18
-	+{static} hexadecimal_string_non_hexadecimal_digit_found  : int = 19
-	+{static} hexadecimal_string_lack_of_greater_than_sign    : int = 20
-	+{static} dictionary_key_is_not_name_object               : int = 21
-	+{static} dictionary_lack_of_double_greater_than_sign     : int = 22
-	+{static} stream_CRLF_or_LF_not_found                     : int = 23
-	+{static} stream_dictionary_absence_of_Length_entry       : int = 24
-	+{static} stream_data_is_shorter_than_Length              : int = 25
-	+{static} keyword_endstream_not_found                     : int = 26
-	+{static} trailer_dictionary_not_found                    : int = 27
-	+{static} indirect_object_is_inconsistent_with_xref_table : int = 28
-	+{static} keyword_obj_not_found                           : int = 29
-	+{static} indirect_object_refers_indirect_reference       : int = 30
-	+{static} keyword_endobj_not_found                        : int = 31
+	+{static} failed_to_seek                                  : int = 1
+	+{static} keyword_EOF_not_found                           : int = 2
+	+{static} xref_byte_offset_not_found                      : int = 3
+	+{static} keyword_startxref_not_found                     : int = 4
+	+{static} EOL_not_found                                   : int = 5
+	+{static} signed_integer_not_found                        : int = 6
+	+{static} unsigned_integer_not_found                      : int = 7
+	+{static} keyword_xref_not_found                          : int = 8
+	+{static} space_not_found                                 : int = 9
+	+{static} xref_entry_first_10_digits_invalid              : int = 10
+	+{static} xref_entry_second_5_digits_invalid              : int = 11
+	+{static} xref_entry_keyword_invalid                      : int = 12
+	+{static} xref_entry_last_2_bytes_invalid                 : int = 13
+	+{static} keyword_trailer_not_found                       : int = 14
+	+{static} literal_string_lack_of_right_parenthesis        : int = 16
+	+{static} unknown_character_detected                      : int = 17
+	+{static} object_not_found                                : int = 18
+	+{static} name_invalid_octal_code                         : int = 19
+	+{static} array_lack_of_right_square_bracket              : int = 20
+	+{static} hexadecimal_string_non_hexadecimal_digit_found  : int = 21
+	+{static} hexadecimal_string_lack_of_greater_than_sign    : int = 22
+	+{static} dictionary_key_is_not_name_object               : int = 23
+	+{static} dictionary_lack_of_double_greater_than_sign     : int = 24
+	+{static} stream_CRLF_or_LF_not_found                     : int = 25
+	+{static} stream_dictionary_absence_of_Length_entry       : int = 26
+	+{static} stream_data_is_shorter_than_Length              : int = 27
+	+{static} keyword_endstream_not_found                     : int = 28
+	+{static} trailer_dictionary_not_found                    : int = 29
+	+{static} indirect_object_is_inconsistent_with_xref_table : int = 30
+	+{static} keyword_obj_not_found                           : int = 31
+	+{static} indirect_object_refers_indirect_reference       : int = 32
+	+{static} keyword_endobj_not_found                        : int = 33
 }
 class parse_error {
 	+enum error_code
