+++ document/stream_parser.class.pu
@@ -53,7 +53,7 @@ stream_parser ..> ignore_flag
 stream_parser ..> xref_types
 stream_parser ..> object_types
 stream_parser <..> object_pool<InputStreamT>
-stream_parser ..> syntax_error
+stream_parser ..> parse_error
 
 annotation xref_types <<(N, yellow) namespace>> {
 	[[xref_types.class.pu]]
@@ -64,7 +64,7 @@ annotation object_types <<(N, yellow) namespace>> {
 class object_pool<InputStreamT> {
 	[[object_pool.class.pu]]
 }
-class syntax_error {
+class parse_error {
 	[[error_types.class.pu]]
 }
 }
