+++ test/UnitTest/PDFParserTest/seek_forward_head_of_line_test.hpp
@@ -1,9 +1,6 @@
 #pragma once
 
-#include "PDFParser.stream_parser.h"
 #include "testtool.h"
-
-#include <sstream>
 
 namespace pdfparser_test {
 [TestClass] public ref class seek_forward_head_of_line_test {
@@ -19,8 +16,5 @@ public:
 	[TestMethod] void test_LF_LFLF();
 	[TestMethod] void test_CRLF_CRCRLF();
 	[TestMethod] void test_CRLF_LFCRLF();
-
-private:
-	pdfparser::stream_parser<std::stringstream>* stream;
 };
 } // namespace pdfparser_test
