+++ test/UnitTest/PDFParserTest/take_any_object@array_object_test.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "testtool.h"
+
+namespace pdfparser_test { namespace take_any_object {
+	[TestClass] public ref class array_object_test {
+	public:
+		// Array Object
+	};
+}} // namespace pdfparser_test::take_any_object
