+++ test/UnitTest/PDFParserTest/take_any_object@hexadecimal_string_test.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "testtool.h"
+
+namespace pdfparser_test { namespace take_any_object {
+	[TestClass] public ref class hexadecimal_string_test {
+	public:
+		// Hexadecimal String
+	};
+}} // namespace pdfparser_test::take_any_object
