+++ test/UnitTest/PDFParserTest/take_any_object@indirect_reference_test.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "testtool.h"
+
+namespace pdfparser_test { namespace take_any_object {
+	[TestClass] public ref class indirect_reference_test {
+	public:
+		// Indirect Reference
+	};
+}} // namespace pdfparser_test::take_any_object
