+++ test/UnitTest/PDFParserTest/take_any_object_test@hexadecimal_string.cpp
@@ -1,0 +1,11 @@
+#include "PDFParser.h"
+#include "take_any_object@hexadecimal_string_test.hpp"
+
+#include <sstream>
+
+using namespace pdfparser;
+using namespace error_types;
+using namespace object_types;
+using namespace pdfparser_test;
+
+// TODO: ŽÀ‘•
