+++ test/UnitTest/PDFParserTest/take_indirect_object_test.cpp
@@ -1,0 +1,11 @@
+#include "PDFParser.h"
+#include "take_indirect_object_test.hpp"
+
+#include <sstream>
+
+using namespace pdfparser;
+using namespace error_types;
+using namespace object_types;
+using namespace pdfparser_test;
+
+// TODO: ƒeƒXƒgì¬
