+++ test/UnitTest/PDFParserTest/take_indirect_object_test.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "testtool.h"
+
+namespace pdfparser_test {
+[TestClass] public ref class take_indirect_object_test {
+public:
+	[TestMethod] void test_();
+};
+} // namespace pdfparser_test
