+++ test/UnitTest/PDFParserTest/take_unsigned_integer_test.hpp
@@ -2,11 +2,8 @@
 
 #include "testtool.h"
 
-#include <sstream>
-
 namespace pdfparser_test {
 [TestClass] public ref class take_unsigned_integer_test {
-public:
 public:
 	[TestMethod] void test_unsigned();
 	[TestMethod] void test_max();
@@ -16,8 +13,5 @@ public:
 	[TestMethod] void test_EOF();
 	[TestMethod] void test_signed();
 	[TestMethod] void test_max_plus1();
-
-private:
-	pdfparser::stream_parser<std::stringstream>* stream;
 };
 } // namespace pdfparser_test
