+++ test/UnitTest/PDFParserTest/tell_test.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "testtool.h"
+
+namespace pdfparser_test {
+[TestClass] public ref class tell_test {
+public:
+	[TestMethod] void test_MIX_EOL();
+};
+} // namespace pdfparser_test
